requirements.txt
aiogram>=3.0
python-dotenv
openai
deepseek ? #можно было бы такую нейросеть использовать, но пока что API в разработке
#нужные зависимости и инструменты